ddl: remove some help function to util_test.go
