+++ ddl/table_test.go
@@ -15,173 +15,20 @@
 package ddl
 
 import (
-	"bytes"
 	"context"
-	"fmt"
 	"testing"
 
-	. "github.com/pingcap/check"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/meta"
 	"github.com/pingcap/tidb/meta/autoid"
-	"github.com/pingcap/tidb/parser/auth"
 	"github.com/pingcap/tidb/parser/model"
-	"github.com/pingcap/tidb/parser/mysql"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/store/mockstore"
 	"github.com/pingcap/tidb/table"
 	"github.com/pingcap/tidb/types"
 	"github.com/stretchr/testify/require"
 )
-
-func testTableInfoWith2IndexOnFirstColumn(c *C, d *ddl, name string, num int) *model.TableInfo {
-	normalInfo, err := testTableInfo(d, name, num)
-	c.Assert(err, IsNil)
-	idxs := make([]*model.IndexInfo, 0, 2)
-	for i := range idxs {
-		idx := &model.IndexInfo{
-			Name:    model.NewCIStr(fmt.Sprintf("i%d", i+1)),
-			State:   model.StatePublic,
-			Columns: []*model.IndexColumn{{Name: model.NewCIStr("c1")}},
-		}
-		idxs = append(idxs, idx)
-	}
-	normalInfo.Indices = idxs
-	normalInfo.Columns[0].FieldType.Flen = 11
-	return normalInfo
-}
-
-// testTableInfo creates a test table with num int columns and with no index.
-func testTableInfo(d *ddl, name string, num int) (*model.TableInfo, error) {
-	tblInfo := &model.TableInfo{
-		Name: model.NewCIStr(name),
-	}
-	genIDs, err := d.genGlobalIDs(1)
-
-	if err != nil {
-		return nil, err
-	}
-	tblInfo.ID = genIDs[0]
-
-	cols := make([]*model.ColumnInfo, num)
-	for i := range cols {
-		col := &model.ColumnInfo{
-			Name:         model.NewCIStr(fmt.Sprintf("c%d", i+1)),
-			Offset:       i,
-			DefaultValue: i + 1,
-			State:        model.StatePublic,
-		}
-
-		col.FieldType = *types.NewFieldType(mysql.TypeLong)
-		col.ID = allocateColumnID(tblInfo)
-		cols[i] = col
-	}
-	tblInfo.Columns = cols
-	tblInfo.Charset = "utf8"
-	tblInfo.Collate = "utf8_bin"
-	return tblInfo, nil
-}
-
-// testTableInfoWithPartition creates a test table with num int columns and with no index.
-func testTableInfoWithPartition(c *C, d *ddl, name string, num int) *model.TableInfo {
-	tblInfo, err := testTableInfo(d, name, num)
-	c.Assert(err, IsNil)
-	genIDs, err := d.genGlobalIDs(1)
-	c.Assert(err, IsNil)
-	pid := genIDs[0]
-	tblInfo.Partition = &model.PartitionInfo{
-		Type:   model.PartitionTypeRange,
-		Expr:   tblInfo.Columns[0].Name.L,
-		Enable: true,
-		Definitions: []model.PartitionDefinition{{
-			ID:       pid,
-			Name:     model.NewCIStr("p0"),
-			LessThan: []string{"maxvalue"},
-		}},
-	}
-
-	return tblInfo
-}
-
-// testTableInfoWithPartitionLessThan creates a test table with num int columns and one partition specified with lessthan.
-func testTableInfoWithPartitionLessThan(c *C, d *ddl, name string, num int, lessthan string) *model.TableInfo {
-	tblInfo := testTableInfoWithPartition(c, d, name, num)
-	tblInfo.Partition.Definitions[0].LessThan = []string{lessthan}
-	return tblInfo
-}
-
-func testAddedNewTablePartitionInfo(c *C, d *ddl, tblInfo *model.TableInfo, partName, lessthan string) *model.PartitionInfo {
-	genIDs, err := d.genGlobalIDs(1)
-	c.Assert(err, IsNil)
-	pid := genIDs[0]
-	// the new added partition should change the partition state to state none at the beginning.
-	return &model.PartitionInfo{
-		Type:   model.PartitionTypeRange,
-		Expr:   tblInfo.Columns[0].Name.L,
-		Enable: true,
-		Definitions: []model.PartitionDefinition{{
-			ID:       pid,
-			Name:     model.NewCIStr(partName),
-			LessThan: []string{lessthan},
-		}},
-	}
-}
-
-// testViewInfo creates a test view with num int columns.
-func testViewInfo(c *C, d *ddl, name string, num int) *model.TableInfo {
-	tblInfo := &model.TableInfo{
-		Name: model.NewCIStr(name),
-	}
-	genIDs, err := d.genGlobalIDs(1)
-	c.Assert(err, IsNil)
-	tblInfo.ID = genIDs[0]
-
-	cols := make([]*model.ColumnInfo, num)
-	viewCols := make([]model.CIStr, num)
-
-	var stmtBuffer bytes.Buffer
-	stmtBuffer.WriteString("SELECT ")
-	for i := range cols {
-		col := &model.ColumnInfo{
-			Name:   model.NewCIStr(fmt.Sprintf("c%d", i+1)),
-			Offset: i,
-			State:  model.StatePublic,
-		}
-
-		col.ID = allocateColumnID(tblInfo)
-		cols[i] = col
-		viewCols[i] = col.Name
-		stmtBuffer.WriteString(cols[i].Name.L + ",")
-	}
-	stmtBuffer.WriteString("1 FROM t")
-
-	view := model.ViewInfo{Cols: viewCols, Security: model.SecurityDefiner, Algorithm: model.AlgorithmMerge,
-		SelectStmt: stmtBuffer.String(), CheckOption: model.CheckOptionCascaded, Definer: &auth.UserIdentity{CurrentUser: true}}
-
-	tblInfo.View = &view
-	tblInfo.Columns = cols
-
-	return tblInfo
-}
-
-func testCreateTable(c *C, ctx sessionctx.Context, d *ddl, dbInfo *model.DBInfo, tblInfo *model.TableInfo) *model.Job {
-	job := &model.Job{
-		SchemaID:   dbInfo.ID,
-		TableID:    tblInfo.ID,
-		Type:       model.ActionCreateTable,
-		BinlogInfo: &model.HistoryInfo{},
-		Args:       []interface{}{tblInfo},
-	}
-	err := d.doDDLJob(ctx, job)
-	c.Assert(err, IsNil)
-
-	v := getSchemaVer(c, ctx)
-	tblInfo.State = model.StatePublic
-	checkHistoryJobArgs(c, ctx, job.ID, &historyJobArgs{ver: v, tbl: tblInfo})
-	tblInfo.State = model.StateNone
-	return job
-}
 
 func testCreateTableT(t *testing.T, ctx sessionctx.Context, d *ddl, dbInfo *model.DBInfo, tblInfo *model.TableInfo) *model.Job {
 	job := &model.Job{
@@ -197,26 +44,6 @@ func testCreateTableT(t *testing.T, ctx sessionctx.Context, d *ddl, dbInfo *mode
 	v := getSchemaVerT(t, ctx)
 	tblInfo.State = model.StatePublic
 	checkHistoryJobArgsT(t, ctx, job.ID, &historyJobArgs{ver: v, tbl: tblInfo})
-	tblInfo.State = model.StateNone
-	return job
-}
-
-func testCreateView(c *C, ctx sessionctx.Context, d *ddl, dbInfo *model.DBInfo, tblInfo *model.TableInfo) *model.Job {
-	job := &model.Job{
-		SchemaID:   dbInfo.ID,
-		TableID:    tblInfo.ID,
-		Type:       model.ActionCreateView,
-		BinlogInfo: &model.HistoryInfo{},
-		Args:       []interface{}{tblInfo, false, 0},
-	}
-
-	c.Assert(tblInfo.IsView(), IsTrue)
-	err := d.doDDLJob(ctx, job)
-	c.Assert(err, IsNil)
-
-	v := getSchemaVer(c, ctx)
-	tblInfo.State = model.StatePublic
-	checkHistoryJobArgs(c, ctx, job.ID, &historyJobArgs{ver: v, tbl: tblInfo})
 	tblInfo.State = model.StateNone
 	return job
 }
@@ -281,21 +108,6 @@ func checkTableLockedTest(t *testing.T, d *ddl, dbInfo *model.DBInfo, tblInfo *m
 	require.NoError(t, err)
 }
 
-func testDropTable(c *C, ctx sessionctx.Context, d *ddl, dbInfo *model.DBInfo, tblInfo *model.TableInfo) *model.Job {
-	job := &model.Job{
-		SchemaID:   dbInfo.ID,
-		TableID:    tblInfo.ID,
-		Type:       model.ActionDropTable,
-		BinlogInfo: &model.HistoryInfo{},
-	}
-	err := d.doDDLJob(ctx, job)
-	c.Assert(err, IsNil)
-
-	v := getSchemaVer(c, ctx)
-	checkHistoryJobArgs(c, ctx, job.ID, &historyJobArgs{ver: v, tbl: tblInfo})
-	return job
-}
-
 func testDropTableT(t *testing.T, ctx sessionctx.Context, d *ddl, dbInfo *model.DBInfo, tblInfo *model.TableInfo) *model.Job {
 	job := &model.Job{
 		SchemaID:   dbInfo.ID,
@@ -331,24 +143,6 @@ func testTruncateTable(t *testing.T, ctx sessionctx.Context, d *ddl, dbInfo *mod
 	return job
 }
 
-func testCheckTableState(c *C, d *ddl, dbInfo *model.DBInfo, tblInfo *model.TableInfo, state model.SchemaState) {
-	err := kv.RunInNewTxn(context.Background(), d.store, false, func(ctx context.Context, txn kv.Transaction) error {
-		t := meta.NewMeta(txn)
-		info, err := t.GetTable(dbInfo.ID, tblInfo.ID)
-		c.Assert(err, IsNil)
-
-		if state == model.StateNone {
-			c.Assert(info, IsNil)
-			return nil
-		}
-
-		c.Assert(info.Name, DeepEquals, tblInfo.Name)
-		c.Assert(info.State, Equals, state)
-		return nil
-	})
-	c.Assert(err, IsNil)
-}
-
 func testCheckTableStateT(t *testing.T, d *ddl, dbInfo *model.DBInfo, tblInfo *model.TableInfo, state model.SchemaState) {
 	err := kv.RunInNewTxn(context.Background(), d.store, false, func(ctx context.Context, txn kv.Transaction) error {
 		tt := meta.NewMeta(txn)
@@ -365,12 +159,6 @@ func testCheckTableStateT(t *testing.T, d *ddl, dbInfo *model.DBInfo, tblInfo *m
 		return nil
 	})
 	require.NoError(t, err)
-}
-
-func testGetTable(c *C, d *ddl, schemaID int64, tableID int64) table.Table {
-	tbl, err := testGetTableWithError(d, schemaID, tableID)
-	c.Assert(err, IsNil)
-	return tbl
 }
 
 func testGetTableT(t *testing.T, d *ddl, schemaID int64, tableID int64) table.Table {
@@ -537,23 +325,4 @@ func testAlterNoCacheTable(t *testing.T, ctx sessionctx.Context, d *ddl, newSche
 	v := getSchemaVerT(t, ctx)
 	checkHistoryJobArgsT(t, ctx, job.ID, &historyJobArgs{ver: v})
 	return job
-}
-
-// for drop indexes
-func createTestTableForDropIndexes(c *C, ctx sessionctx.Context, d *ddl, dbInfo *model.DBInfo, name string, num int) *model.TableInfo {
-	tableInfo, err := testTableInfo(d, name, num)
-	c.Assert(err, IsNil)
-	var idxs []*model.IndexInfo
-	for i := 0; i < num; i++ {
-		idxName := model.NewCIStr(fmt.Sprintf("i%d", i+1))
-		idx := &model.IndexInfo{
-			Name:    idxName,
-			State:   model.StatePublic,
-			Columns: []*model.IndexColumn{{Name: model.NewCIStr(fmt.Sprintf("c%d", i+1))}},
-		}
-		idxs = append(idxs, idx)
-	}
-	tableInfo.Indices = idxs
-	testCreateTable(c, ctx, d, dbInfo, tableInfo)
-	return tableInfo
 }
